Fix some more warnings

+++ src/boss/code_0.c
@@ -362,12 +362,12 @@ void func_boss_800253AC(void) {
     func_global_asm_807149B8(1);
     func_global_asm_8071498C(&func_global_asm_8071AF30);
     func_global_asm_807149A8(1000);
-    func_global_asm_80714950(current_actor_pointer);
+    func_global_asm_80714950((s32)current_actor_pointer);
     func_global_asm_80714C08(&D_global_asm_80720120, scale, current_actor_pointer, 3, 0);
     func_global_asm_807149B8(1);
     func_global_asm_8071498C(&func_global_asm_8071AF30);
     func_global_asm_807149A8(1000);
-    func_global_asm_80714950(current_actor_pointer);
+    func_global_asm_80714950((s32)current_actor_pointer);
     func_global_asm_80714C08(&D_global_asm_80720120, scale, current_actor_pointer, 4, 0);
 }
